:tractor: Make AddFieldMiddleware a RowMiddleware
