+++ pkg/middlewares/row/add-field.go
@@ -1,0 +1,26 @@
+package row
+
+import (
+	"context"
+	"github.com/go-go-golems/glazed/pkg/types"
+)
+
+type AddFieldMiddleware struct {
+	Fields map[string]string
+}
+
+func NewAddFieldMiddleware(fields map[string]string) *AddFieldMiddleware {
+	return &AddFieldMiddleware{Fields: fields}
+}
+
+func (a *AddFieldMiddleware) Process(ctx context.Context, row types.Row) ([]types.Row, error) {
+	newValues := types.NewMapRow()
+	for pair := row.Oldest(); pair != nil; pair = pair.Next() {
+		key, value := pair.Key, pair.Value
+		newValues.Set(key, value)
+	}
+	for key, value := range a.Fields {
+		newValues.Set(key, value)
+	}
+	return []types.Row{newValues}, nil
+}
