+++ pkg/settings/settings_replace.go
@@ -5,6 +5,7 @@ import (
 	"github.com/go-go-golems/glazed/pkg/cmds/layers"
 	"github.com/go-go-golems/glazed/pkg/cmds/parameters"
 	"github.com/go-go-golems/glazed/pkg/middlewares"
+	"github.com/go-go-golems/glazed/pkg/middlewares/row"
 	"github.com/go-go-golems/glazed/pkg/middlewares/table"
 	"github.com/pkg/errors"
 	"os"
@@ -31,8 +32,8 @@ func (rs *ReplaceSettings) AddMiddlewares(of *middlewares.Processor) error {
 	}
 
 	if len(rs.AddFields) > 0 {
-		mw := table.NewAddFieldMiddleware(rs.AddFields)
-		of.AddTableMiddleware(mw)
+		mw := row.NewAddFieldMiddleware(rs.AddFields)
+		of.AddRowMiddleware(mw)
 	}
 
 	return nil
